Provide mechanism for unprivileged caching
By default, omnibus wants to cache to /var/cache/omnibus/*, therefore
set each of the cache directories to be local or user defined. This
may seem weird that we set each individually, but omnibus does not
allow for setting a cache base directory. I dont want to have to rely
on a pull request, so just set all four directories.