Complete rework of the build process
Now we require a manifest file that will tell omnibus-openstack
exactly which projects you care about. If its in the manifest, it
makes it, and if it's not it doesn't.  That manifest may be json or
yaml.

Each project will now be a separate venv within the resulting package.

One package to rule them all.