Alter the way we build openstack projects
Previously we had relied on a fork of upstream to support naming a
software object. This obviously meant that the upstream code could
change and leave us in the dust. So, in order to combat that problem
I have changed to using this envvar hack to get the software name
exposed to the software object early in the constructor.

More context:
https://github.com/opscode/omnibus-ruby/pull/123